+++ include/Mesh.h
@@ -12,7 +12,7 @@ private:
 
     const std::string _name;
     const std::vector<SGEPosition> _vertices;
-    const std::vector<uint16_t> _indices;
+    const std::vector<uint32_t> _indices;
 	const bool _filled;
 	bool _textured;
 
@@ -25,9 +25,9 @@ private:
 
 	// default meshes (vertices with indices)
 	static const std::vector <SGEPosition> defaultRectangleVertices;
-	static const std::vector<uint16_t> defaultRectangleIndices;
+	static const std::vector<uint32_t> defaultRectangleIndices;
 	static const std::vector <SGEPosition> defaultTriangleVertices;
-	static const std::vector<uint16_t> defaultTriangleIndices;
+	static const std::vector<uint32_t> defaultTriangleIndices;
 
 protected:
 
@@ -48,7 +48,7 @@ public:
 
 	void setTextured(bool isTextured) { _textured = isTextured; }
 
-    Mesh(const std::string name, const std::vector<SGEPosition> vertices, const std::vector<uint16_t> indices, const bool filled = true);
+    Mesh(const std::string name, const std::vector<SGEPosition> vertices, const std::vector<uint32_t> indices, const bool filled = true);
 
 	static Mesh getDefaultRectangleMesh(const std::string name, const bool filld = true);
 	static Mesh getDefaultTriangleMesh(const std::string name, const bool filld = true);
